Changed repositories and controllers implementation

+++ CorepetitorApi/Controllers/ModulesController.cs
@@ -3,7 +3,7 @@ using CorepetitorApi.Repositories;
 using CorepetitorApi.Models;
 using System.Collections.Generic;
 
-[Route("api/[controller]")]
+[Route("api/tutors/{tutorId}/[controller]")]
 [ApiController]
 public class ModulesController : ControllerBase
 {
@@ -14,41 +14,48 @@ public class ModulesController : ControllerBase
         _repository = repository;
     }
 
-    // GET: api/Modules
+    // GET: api/tutors/{tutorId}/Modules
     [HttpGet]
-    public ActionResult<IEnumerable<Module>> GetAll() => Ok(_repository.GetAllModules());
-
-    // GET: api/Modules/{id}
-    [HttpGet("{id}")]
-    public ActionResult<Module> Get(int id)
+    public ActionResult<IEnumerable<Module>> GetAll(int tutorId)
     {
-        var module = _repository.GetModuleById(id);
+        return Ok(_repository.GetAllModules(tutorId));
+    }
+
+    // GET: api/tutors/{tutorId}/Modules/{id}
+    [HttpGet("{id}")]
+    public ActionResult<Module> Get(int tutorId, int id)
+    {
+        var module = _repository.GetModuleById(tutorId, id);
         if (module == null) return NotFound();
         return Ok(module);
     }
 
-    // POST: api/Modules
+    // POST: api/tutors/{tutorId}/Modules
     [HttpPost]
-    public ActionResult<Module> Add(Module module)
+    public ActionResult<Module> Add(int tutorId, Module module)
     {
-        _repository.AddModule(module);
-        return CreatedAtAction(nameof(Get), new { id = module.Id }, module);
+        _repository.AddModule(tutorId, module);
+        return CreatedAtAction(nameof(Get), new { tutorId, id = module.Id }, module);
     }
 
-    // PUT: api/Modules/{id}
+    // PUT: api/tutors/{tutorId}/Modules/{id}
     [HttpPut("{id}")]
-    public ActionResult Update(int id, Module module)
+    public ActionResult Update(int tutorId, int id, Module module)
     {
         if (id != module.Id) return BadRequest();
 
-        _repository.UpdateModule(module);
+        var updatedModule = _repository.UpdateModule(tutorId, module);
+        if (updatedModule == null) return NotFound();
+
         return NoContent();
     }
 
-    // DELETE: api/Modules/{id}
+    // DELETE: api/tutors/{tutorId}/Modules/{id}
     [HttpDelete("{id}")]
-    public ActionResult<Module> Delete(int id)
+    public ActionResult<Module> Delete(int tutorId, int id)
     {
+        if (!_repository.ModuleExists(id, tutorId)) return NotFound();
+
         _repository.DeleteModule(id);
         return NoContent();
     }
